Several improvements.
Several improvements have been made to make the API more consistent. Some documentation has been modified. But there is still a bug that causes the training script to not detect the end of an episode when the Time.timeScale variable is greater than 1. Training with stable-baselines3 stopped converging in this release. We are reviewing the code to resolve this issue.