Several improvements was applied in data shape convention.
Several improvements was applied in data shape convention: now, all a shape is structured as a tuple (A1, A2, ..., AN) and one is defined in the specific type of the sensor. Non-necessary exported fields was take out of the  Sensor in Sensor.cs. Specific sensors, if convenient, can export extra fields.